+++ OVHJabberBot/commands/__init__.py
@@ -1,0 +1,5 @@
+import pkgutil
+
+__path__ = pkgutil.extend_path(__path__, __name__)
+for importer, modname, ispkg in pkgutil.walk_packages(path=__path__, prefix=__name__+'.'):
+      __import__(modname)
