+++ OVHJabberBot/db/insulte.py
@@ -1,0 +1,5 @@
+from mongoengine import *
+
+class Insulte(Document):
+    meta = {'collection': 'insultes'}
+    text = StringField(required=True)
