+++ OVHJabberBot/db/user.py
