--- db/__init__.py
@@ -1,3 +1,0 @@
-from mongoengine import connect
-
-connect("boulanger")
